refactored tinyfatdb folder into separate files
